+++ client/src/components/manageProjects/editMeetingTimes.js
@@ -3,12 +3,15 @@ import '../../sass/ManageProjects.scss';
 import EditableMeeting from './editableMeeting';
 import { REACT_APP_CUSTOM_REQUEST_HEADER } from "../../utils/globalSettings";
 import CreateNewEvent from './createNewEvent';
+import { readableEvent } from './utilities/readableEvent';
+import { findNextOccuranceOfDay } from './utilities/findNextDayOccuranceOfDay';
+import { addDurationToTime } from './utilities/addDurationToTime';
+import { timeConvertFromForm } from './utilities/timeConvertFromForm';
 
 // This component displays current meeting times for selected project and offers the option to edit those times. 
 const EditMeetingTimes  = ({
   recurringEvents, 
   projectToEdit,
-  fetchRecurringEvents,
   goEditProject,
   goSelectProject
   }) => {
@@ -16,10 +19,7 @@ const EditMeetingTimes  = ({
   const headerToSend = REACT_APP_CUSTOM_REQUEST_HEADER;
   const URL = process.env.NODE_ENV === 'prod' ? 'https://www.vrms.io' : 'http://localhost:4000';
 
-  // Initialize state
   const [rEvents, setREvents] = useState([]);
-
-  // ToDo: Is this state needed? If not, remove it. 
   const [eventToEdit, setEventToEdit] = useState('');
 
   // Get project recurring events when component loads
@@ -27,50 +27,6 @@ const EditMeetingTimes  = ({
     // Filters the recurring events for this project
     setREvents(recurringEvents.filter(e => (e?.project?._id === projectToEdit._id)));
   }, [])
-
-  // Translate event data into human readable format
-  function readableEvent (e) {
-
-    // Get date for each of the parts of the event time/day   
-    let d = new Date(e.date);
-    let start = new Date(e.startTime);
-    let end = new Date(e.endTime);
-
-    //Get day of the week. (Get the day number for sorting)
-    let options = { weekday: "long" };
-    let dayOfTheWeek = Intl.DateTimeFormat("en-US", options).format(d);
-    let dayOfTheWeekNumber = d.getDay();
-
-    // Convert end time from 24 to 12 and make pretty
-    let sHours = start.getHours();
-    let startHours = (sHours % 12) || 12;
-    let startMinutes = (start.getMinutes() < 10 ? '0' : '') + start.getMinutes();
-    let startAorP = sHours >= 12 ? 'pm' : 'am';
-    let startTime = startHours + ":" + startMinutes + startAorP;
-
-    // Convert end time from 24 to 12 and make pretty
-    let eHours = end.getHours();
-    let endHours = (eHours % 12) || 12;
-    let endMinutes = (end.getMinutes() < 10 ? '0' : '') + end.getMinutes();
-    let endAorP = eHours >= 12 ? 'pm' : 'am';
-    let endTime = endHours + ":" + endMinutes + endAorP;
-
-    // Create readable object for this event
-    let newEvent = {
-      name: e.name,
-      description: e.description,
-      eventType: e.eventType,
-      dayOfTheWeekNumber: dayOfTheWeekNumber,
-      dayOfTheWeek: dayOfTheWeek,
-      startTime: startTime,
-      endTime: endTime,
-      duration: e.hours,
-      event_id: e._id,
-      videoConferenceLink: e.videoConferenceLink
-    }
-
-    return newEvent;
-  }
 
   // Map new array of readable event objects
   let processedEvents = rEvents.map(function(item) {
@@ -90,7 +46,7 @@ const EditMeetingTimes  = ({
     let uEvents = [];
     
     for (let i = 0; i < rEvents.length; i++) {
-      if (rEvents[[i]]._id === data._id) {
+      if (rEvents[i]._id === data._id) {
         uEvents.push(data);
       } else {
         uEvents.push(rEvents[i]);
@@ -98,8 +54,6 @@ const EditMeetingTimes  = ({
     }
     setREvents(uEvents);
   };
-
-    /*** Update Event Functions ***/
 
   // update reurringEvent
   const updateRecurringEvent = async (eventToUpdate, RecurringEventID) => {
@@ -125,32 +79,32 @@ const EditMeetingTimes  = ({
 
     setEventToEdit(eventID);
 
-    // ToDo: Add some validation if you haven't made any changes, or have deleted the description field
+    // ToDo: Add some validation.  What kind of validation do we need?
+    // ToDo: Add some update confirmation so user knows the update was successful
 
-    // Values that have been changed for will be added to the object and, in time, the database
     let theUpdatedEvent = {};
 
-    // If the fields have been changed, add to object
-
-    // Fields that don't need any processing
     if (values.name) {
       theUpdatedEvent = { 
         ...theUpdatedEvent, 
         name: values.name 
       };
     }
+
     if (values.eventType) {
       theUpdatedEvent = { 
         ...theUpdatedEvent, 
         eventType: values.eventType 
       };
     }
+
     if (values.description) {
       theUpdatedEvent = { 
         ...theUpdatedEvent, 
         description: values.description 
       };
     }
+
     if (values.meetingURL) {
       theUpdatedEvent = { 
         ...theUpdatedEvent, 
@@ -165,14 +119,9 @@ const EditMeetingTimes  = ({
       updatedDate: updatedDate 
     };
 
-    // These need some processing before adding to the update object
-
     // If the day has been changed, find the next occurence of the changed day
     if (values.dayNumber) {
-      let day = parseInt(values.dayNumber);
-      const date = new Date();
-      date.setDate(date.getDate() + ((7 - date.getDay()) % 7 + day) % 7); 
-      
+      const date = findNextOccuranceOfDay(values.dayNumber);
       const dateGMT = new Date(date).toISOString();
 
       theUpdatedEvent = {
@@ -183,10 +132,6 @@ const EditMeetingTimes  = ({
 
     // Set start time, End time and Duration if either start time or duration is changed
     if ( values.startTime || values.duration ) {
-
-      // ToTo: A lot of this code is reused from CreateNewEvent.js and could possibly be consolidated
-      // Yes, I know, I should do it now, but I just want to get it working and then
-      // I'll come back to it.  (See "Technical Debt")
 
       /*
       We need a start time and a duration to calculate everything we need.  
@@ -205,66 +150,8 @@ const EditMeetingTimes  = ({
         durationToUse = values.duration;
       }
 
-      const timeDate = new Date();
-
-      // ToDo: Make this a reusable function
-      // reconstitute time from form back into timestamp
-      let timeParts = startTimeToUse.split(':');
-      const sap = timeParts[1].slice(-2);
-      timeParts[1] = timeParts[1].slice(0,-2);
-      let startHour = parseInt(timeParts[0]);
-      const startMinutes = parseInt(timeParts[1]);
-      const startSeconds = 0;
-      
-      // set 12am to 0 and make afternoon into military time
-      if (sap === 'pm' && startHour !== 12) {
-        startHour = startHour + 12;
-      } else if (sap === 'am' && startHour === 12) {
-        startHour = 0;
-      }
-  
-      // Update the date string with the start hours of the meeting
-      timeDate.setHours(startHour);
-      timeDate.setMinutes(startMinutes);
-      timeDate.setSeconds(startSeconds);
-  
-      // This is the date and time of the first meeting.
-      // This will also be used as the start time
-      //const startTimeDate = new Date(date.getTime());
-
-      // ToDo: Make this a reusable function
-      let endTime;
-
-      // Create the endTime by adding seconds to the timestamp and converting it back date
-      switch (durationToUse) {
-        case '.5':
-          endTime = new Date(timeDate.getTime() + (.5*3600000)); 
-        break;
-        case '1':
-          endTime = new Date(timeDate.getTime() + (1*3600000)); 
-        break;
-        case '1.5':
-          endTime = new Date(timeDate.getTime() + (1.5*3600000)); 
-        break;
-        case '2':
-          endTime = new Date(timeDate.getTime() + (2*3600000)); 
-        break;
-        case '2.5':
-          endTime = new Date(timeDate.getTime() + (2.5*3600000));
-        break;
-        case '3':
-          endTime = new Date(timeDate.getTime() + (3*3600000));
-        break;
-        case '3.5':
-          endTime = new Date(timeDate.getTime() + (3.5*3600000));
-        break;
-        case '4':
-          endTime = new Date(timeDate.getTime() + (4*3600000));
-        break;
-        default:
-          // I can't think of how it will get to default,  but I thought I'd put this here anyway
-          endTime = new Date(timeDate.getTime()) 
-      } 
+      const timeDate = timeConvertFromForm(new Date(), startTimeToUse);
+      const endTime = addDurationToTime(timeDate, durationToUse); 
 
       //convert to ISO and GMT
       const startTimeGMT = new Date(timeDate).toISOString();
@@ -290,9 +177,7 @@ const EditMeetingTimes  = ({
   }
 
   /*** Delete Event Functions ***/
-
   const deleteRecurringEvent = async (RecurringEventID) => {
-
     const url = `/api/recurringEvents/${RecurringEventID}`;
     const requestOptions = {
         method: 'DELETE',
@@ -308,10 +193,10 @@ const EditMeetingTimes  = ({
     }
     const data = await response.json();
     return data;
-
   }
 
   const handleEventDelete = (eventID) => () => {
+    // ToDo: Add delete confirmation so user knows the item has been deleted
 
     deleteRecurringEvent(eventID)
     .then( (data) => {
@@ -358,5 +243,4 @@ const EditMeetingTimes  = ({
   );
 
 };
-
 export default EditMeetingTimes;
